fix: resolve C++ compile and linker errors for debug APK build

+++ Projects/oblivion-android/app/src/main/cpp/game/quest_manager.cpp
@@ -312,7 +312,7 @@ void QuestManager::logQuestStatus() const {
         auto quest = getQuest(questId);
         if (quest) {
             LOGD("  Quest: %s (ID=%u, Objectives=%zu/%zu)", quest->title.c_str(),
-                 questId, quest->getCompletedObjectiveCount(), quest->objectives.size());
+                 questId, static_cast<size_t>(quest->getCompletedObjectiveCount()), quest->objectives.size());
         }
     }
